combined qaoa solver with multiprocess, with npr 1 being only one processor
